Replace window event handlers with document ones for global hotkeys
useGlobalHotkey is greatly simplified to not use a pool of handlers in order to avoid setting up multiple keydown event handlers on the document. This is due to the fact that the addEventListener API is really fast and typically there will be no more than a couple of global hotkeys set up at a time.