Add model binding
Allows a model hook to accept a binder callback as as second argument. It is called whenever changes on the model occur and it is allowed to handle them and update the component tree instead of the model itself.

This allows for a performance improvement when child models have to be passed on to a parent, while the parent updates this child component whenever its data is passed back up. It skips an intermediary update of the child.

In order to achieve this state is now stored in a ref object instead of a reducer and updates are forcefully called whenever necessary.

- Adds tests to confirm that the model updates correctly.
- modelHookReducer's logic has been moved into the useModel's file in order to be more tightly coupled