Model $errors method dispatches validation action
model.$errors() method now dispatches a validation action and will cause a re-render, if a model value is considered to have changed since the last validation. Otherwise it will only return the list of errors.

This way the $errors() and $firstError() can be called repeatedly without the previous performance impact due to re-validation on every single call.