Remove onBlur and add $validate to Model hook
onBlur was pretty much redundant and is replaced by model.$validate, which can be used in the same way, just a little more explicitly. This makes spreading object entries a lot safer and simpler in many cases.

Also due to the fact that a new function was created for each model entry, it also reduces allocation by a tiny little bit.