Add status and middleware to useTask hook
> It is expected that there are NO BREAKING changes

Task hook now has a status:
 - `null` when it is initialized or cleared
 - `running` while it is running a task
 - `done` when a task has finished successfully
 - `error` when a task run has failed

The previously exposed `isRunning` property remains unchanged for backwards compatibility and ease of use.

Middleware can now be applied to each task. Middleware functions are provided with the running task Promise and can await it to receive its result or undefined value if cancelled. They allow to modify the task execution in a variety of ways, such as ensuring that task runs a minimum amount of time.

Last but not least Task hook is now fully tested!