[Breaking] Add useInputValue hook
## BREAKING
- model.$firstError() and model.$errors() no longer cause a component render. Fixed by using $validate in place of $errors or a combination of $validate and $firstError in place of $firstError

useModel and useInputValue are now inside the same file, called InputHooks.

useInputValue currently only allows a single value to be modeled with the default input parser. In the future it will be possible to "map" it to a model instance, which will be updated passively, preventing updates on all components using the model instance, and instead only updating the component with useInputValue hook, mapped to the model instance. It is currently implemented but it is unstable until after the tests have been written.