[Breaking] Don't call useModel custom paresers on init
Custom parsers should work like "input reducers" and thus should not be called on initialization since there is no input to reduce. Calling them with input as `undefined` makes the parsers' implementation hard in many cases and sometimes leads to unexpected bugs.

On the other hand value initialization for the entry with parser has no place inside the parser function. It should therefore be performed in place when defining the model structure.