Cache $data() result in useModel
Model hook caches the results of $data() method in order to improve performance of subsequent calls.

Also made some internal house-keeping wherever `lastest.current` is called more than once. It is saved as a variable with name `model` in order to make it clear that this is THE model to be used in this method/function and also save some bundle bytes.