Add buildUrlQuery web helper
Works with arrays, similar to how PHP's build_url_query() works. For example array `test = [1, 2]` will be converted to `test[]=1&test[]=2` which is then parsed by PHP to an array. The previously available helper createQuery() would convert this same array to `test=1,2`, which will be parsed by PHP as a string `1,2`.

Both query building functions are currently available, but in the future this behaviour could be merged into a single functions with a toggle option between the two.

request() helper is now using thes new buildUrlQuery helper. This could be a BREAKING change in some cases.