[input-hooks] Add model $submit and serialization
$submit
---------

$submit() is a new method defined on the model object, which when called performs model validation and if there are no errors, calls a provided submission callback with the model's data.

Submission is typically not performed, if  the modelled data is considered to be invalid. That is why this method handles this by default.

Also in almost all cases it is undesired that the user is able to perform submission while still waiting for one by, for example, pressing the button twice, thrice or more. In such a case a faulty records and duplicates could easily end up being stored. This is  the second thing addressed by this method, which discards any subsequent call, until the first one has finished waiting for the asynchronous submission.

serialization
--------------

Custom serialization functions are defined as part of  the initial model structure. They allows for final modifications and transformations of each value, when being retrieved as a data object from the model.

This way expensive or otherwise unnecessary transformations to the modelled value can be applied, before submitting it to the server.

> Note that validation is still performed on the original modelled value, before serialization.