Add useProgressiveLoader hook
This hook provides functionality for progressive loading. Meaning that it progressively fires a loading callback and adds its resulting items to an internal list, until there is nothing more to be loaded. At this point a `finished` state variable is provided to indicate that the progressive loader has "expired". When this happens any calls to the loader callback will be discarded.

Progressive loading is useful for infinite scrollers, news feeds, and other types of lists, which are progressively loaded one piece at a time.

-----

An example use case is a blog, which has some posts initially loaded and a `Load more` button at the end, which progressively loads more and more posts, anytime it is pressed, until there are no more posts to be loaded.