[InputHooks] Allow a single call to $submit
Model's $submit method can now be called only once, unless the model changes.

This method is intended to be used for data submission to the server for example, in which case submitting the same data twice is undesired, and can easily lead to duplicate records.

Furthermore, blocking following calls to the $submit method, after the first one has succeeded and the model hasn't changed, allows for model autosaving to be implemented via React's useEffect hook by providing the model object itself as a depenency.