[InputHooks] Unify internal state
Internal state such as _data and _errors cache, submitted status as well as the latest model object are combined into a single state object. This way the internal state is easier and more straightforward to follow and maintain.

Also the _data and _errors cache were previously exposed, although 'hidden' by TypeScript, to the public, which is no longer the case, and they are encapsulated. This should also reduce the risk of memory leaks, when a model object, along with its cache is stored somewhere, and not cleaned up.