[useModel] Model as second argument for validate
Model entries' validate function receives the latest model instance as a second argument, instead of the model's serialized data object.

This way a value can be validated in place reliably save rare cases when multiple model values rely on each other for validation and a batch change is dispatched via the model's $change() method.

Previously the 'quick' validation whenever a change occured was only attempted. It would fail anytime it relied on the model's data object, which was not provided in this case in order to not impair performance.