Introduce Transfer package
> Inspired by Relay and aims to provide a much simplified experience for data fetching from practically any source.

It builds on top of the existing `createFetcher()` in order to save time and provide a fully backwards compatible new opt-in features.

This package aims to improve the existing Suspense data fetching experience by providing a new hook `useResource`, which reads the resource data and subscribes the component to resource updates. Any time the resource's internal data is changed the subscribed component will rerender with the new data.

Two new methods have been introduced on the Fetcher object:
  - `refresh()` -> fetches new resource data from the endpoint
  - `mutate()` -> allows a resource mutation to be performed and then refreshes the resource and all other caches inside the Fetcher's `options.refreshes` array.

Furthermore any changes made by a Fetcher method to a cached resource will notify the resource's subscribers.