[Transfer] Major invalidation improvements
Previsouly called refresh now the method is called invalidate.

It used to fetch data regardless of whether the resource has any subscribers or not, which resulted in unnecessary requests.

Now it only ever fetches data if the resource is present in cache and it has subscribed components. Otherwise clears cache or does nothing at all. Also it now properly handles request errors, which were previously not handled at all.

Perhaps an "eager" option has to be added in order to optionally make invalidation work like it used to do, prior to this commit.